refactor: replace write and read to another folder
