+++ src/main/kotlin/io/Read.kt
@@ -1,10 +1,13 @@
-package graph
+package io
 
+import graph.DirectedGraph
+import graph.Graph
+import graph.UndirectedGraph
 import java.io.BufferedReader
 import java.io.FileReader
 
 class Read(name: String) {
-    private val reader = BufferedReader(FileReader("$name.csv"))
+    private val reader = BufferedReader(FileReader(name))
 
     fun getGraph(): Graph<String> {
         val graph = readData()
